reorganization of chunk benchmark  #2269 #2672
